fixed mongodb config

+++ src/main/java/com/fypgrading/reviewservice/repository/EvaluationRepository.java
@@ -2,10 +2,11 @@ package com.fypgrading.reviewservice.repository;
 
 import com.fypgrading.reviewservice.entity.Evaluation;
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface EvaluationRepository extends MongoRepository<Evaluation, Integer> {
+public interface EvaluationRepository extends MongoRepository<Evaluation, Long> {
 
     Long countByTeamId(Integer teamId);
 }
